refactor: rename the app name
