+++ src/main/resources/application-dev.properties
@@ -5,7 +5,7 @@ spring.profiles.active=dev
 
 # Logging configuration for development
 logging.level.root=INFO
-logging.level.com.iam.event_cast=INFO
+logging.level.com.iam.metaping=INFO
 
 # Console output for easier development
 spring.output.ansi.enabled=ALWAYS
