+++ src/main/resources/application-prod.properties
@@ -5,7 +5,7 @@ spring.profiles.active=prod
 
 # Logging configuration for production
 logging.level.root=WARN
-logging.level.com.iam.event_cast=INFO
+logging.level.com.iam.metaping=INFO
 logging.level.org.springframework=WARN
 
 # Disable development features in production
