+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
-spring.application.name=event-cast
+spring.application.name=meta-ping
 
 #Spring Cloud Function requires a proper configuration for function registration
 spring.cloud.function.definition=reverseString
 
 # Logging configuration
 logging.level.root=INFO
-logging.level.com.iam.event_cast=DEBUG
+logging.level.com.iam.metaping=DEBUG
